Add a test
